requirements.txt
PyYAML
tqdmt